Separated Ride Request creation into many fragments
