+++ app/src/main/res/layout/fragment_register3.xml
@@ -5,8 +5,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    android:paddingLeft="20dp"
-    android:paddingRight="20dp"
+    android:paddingLeft="30dp"
+    android:paddingRight="30dp"
     tools:context=".SignUp.RegisterFragment2">
 
     <TextView
@@ -66,7 +66,8 @@
         android:background="@drawable/rounded_corners"
         android:text="Finish"
         android:textColor="#FFFFFF"
-        android:textStyle="bold" />
+        android:textStyle="bold"
+        android:visibility="gone" />
 
 
 </LinearLayout>
